move offsets parsing code to its own file
The offsets parsing code is self-contained and complex enough
that it seems worth having in its own file.

This is a pure code move, no other changes.
